+++ core/rust-lib/src/lib.rs
@@ -241,6 +241,7 @@ pub fn run(context: tauri::Context<Wry>) {
             commands::system_reboot,
             commands::system_shutdown,
             commands::system_lock,
+            commands::adjust_volume,
             commands::get_theme_preference,
             commands::set_theme_preference,
             commands::get_screen_recording_status,